+++ app/page.tsx
@@ -2,13 +2,19 @@
 
 import { PiListFill } from 'react-icons/pi';
 import { BsFillGridFill } from 'react-icons/bs';
-import globalContext from '@/context/context';
-import { useContext } from 'react';
-import Filters from '@/components/Filters';
+import DesktopFilter from '@/components/FIlters/DesktopFilter';
 import ImmobileList from '@/components/ImmobileList';
 import MostUsedFilters from '@/components/MostUsedFilters';
+import OverlaidFilter from '@/components/FIlters/OverlaidFilter';
+import { useState } from 'react';
 
 export default function Home() {
+  const [openOverlaidFilter, setOpenOverlaidFilter] = useState(false);
+
+  function handleOpenOverlaidFilter(){
+    setOpenOverlaidFilter(true);
+  }
+
   return (
     <main className='flex flex-col gap-4 items-center w-full h-[calc(100vh-80px)] 
     min-[700px]:pl-20 min-[700px]:flex-row min-[700px]:gap-0 min-[700px]:items-stretch'>
@@ -18,15 +24,19 @@ export default function Home() {
           <div className='flex  items-center gap-2 pr-2 text-[#424242] text-2xl'>
             <PiListFill className='hover:text-zinc-500 cursor-pointer transition-all' />
             <BsFillGridFill className='hover:text-zinc-500 cursor-pointer transition-all' />
-            <button className='border-[#C1C1C1] border-l
-            text-center font-medium text-[#424242] text-base px-5 min-[1466px]:hidden h-full'>
+            <button 
+              className='border-[#C1C1C1] border-l text-center font-medium text-[#424242] text-base px-5 h-full
+              min-[1466px]:hidden '
+              onClick={handleOpenOverlaidFilter}  
+            >
               Filtros
             </button>
           </div>
         </div>
         <ImmobileList />
       </div>
-      <Filters />
+      <DesktopFilter />
+      <OverlaidFilter openOverlaidFilter={openOverlaidFilter} closeOverlaidFilter={setOpenOverlaidFilter}/>
     </main>
   );
 }
